refactor(cf): relocate cf app-detail signal configs out of list framework
