+++ src/frontend/packages/cloud-foundry/src/features/applications/application/application-tabs-base/tabs/variables-tab/variables-tab.component.ts
@@ -29,7 +29,7 @@ import {
 } from '../../../../../../shared/components/list/list-types/app-variables/cf-app-variables.types';
 import {
   CfAppVariablesSignalConfigService,
-} from '../../../../../../shared/components/list/list-types/app-variables/cf-app-variables-signal-config.service';
+} from '../../../../../../shared/signal-list-configs/app-variables/cf-app-variables-signal-config.service';
 import { AppVariableActionsService } from '../../../../../../shared/services/app-variable-actions.service';
 import { AppDetailDataService } from '../../../../app-detail-data.service';
